mobilenet_v2_animals/requirements.txt
paddlepaddle
paddlehub
Pillow
numpy